python-app/requirements.txt
fastapi==0.98.0
uvicorn[standard]==0.22.0
httpx==0.24.1
aio-pika==8.3.0
pydantic==1.10.12
python-dotenv==1.1.0
tenacity==8.2.2